+++ src/modules.rs
@@ -1,18 +1,22 @@
 /// Contains modules for working with journal log files providing readers for individual files or
 /// all files at once.
+#[cfg(feature = "logs")]
 pub mod logs;
 
 /// Allows for tracking journal directory as a whole, firing events for when logs change but also
 /// for when `.json` files are updated.
+#[cfg(feature = "journal")]
 pub mod journal;
 
 /// Utilities for working with the `Status.json` file. The `Status.json` file contains information
 /// about the current status of the player and their ship. Things like if the player is docked
 /// or landed or the location of the player on a planet.
+#[cfg(feature = "json-models")]
 pub mod status;
 
 /// Used to construct a state by feeding it events from the journal logs which can then be used
 /// to create relations between events and create a history.
+#[cfg(feature = "state")]
 pub mod state;
 
 /// Contains models for genuses, species and variants and utilities for predicting possible
@@ -26,28 +30,34 @@ pub mod exploration;
 /// Utilities for working with the `Outfitting.json` file. Is updated when opening the outfitting
 /// menu on stations and includes all the modules that are available for purchase at that market.
 /// Is reset every time the player opens the outfitting menu.
+#[cfg(feature = "json-models")]
 pub mod outfitting;
 
 /// Utilities for working with the `Shipyard.json` file. Is updated when opening the shipyard menu
 /// on stations and lists all the ships that are available to buy at that location. Resets every
 /// time the player opens the shipyard menu.
+#[cfg(feature = "json-models")]
 pub mod shipyard;
 
 /// Utilities for working with the `Market.json` file. Includes a list of commodities that are being
 /// sold/requested at the current location. Resets every time the player opens the commodity market.
+#[cfg(feature = "json-models")]
 pub mod market;
 
 /// Utilities for working with the `NavRoute.json` file. Provides the current plotted nav route for
 /// the player. Resets if the player selects a new route destination or when clearing the route.
+#[cfg(feature = "json-models")]
 pub mod nav_route;
 
 /// Utilities for working with the `ModulesInfo.json` file. This file provides information about
 /// the current state of the modules of the player's ship.
+#[cfg(feature = "json-models")]
 pub mod modules_info;
 
 /// Utilities for working with the `Backpack.json` file. Contains all the items that the player
 /// currently has in their backpack. This is only filled if the player has the Odyssey expansion.
 /// Updates when there are changes to the player's backpack.
+#[cfg(feature = "json-models")]
 pub mod backpack;
 
 /// Utilities for working with the `Cargo.json` file. Includes a list of all the commodities that
@@ -55,10 +65,12 @@ pub mod backpack;
 /// an SRV, so at one time it might show the inventory of the player's ship and the next it shows
 /// the inventory of the SRV, but never both at once. Is updated when there are changes to the
 /// cargo hold or when switching between ship and SRV and vise versa.
+#[cfg(feature = "json-models")]
 pub mod cargo;
 
 /// Utilities for working with the `ShipLocker.json` file. This file includes all the Odyssey items
 /// the player has stored on their ship. Is updated when there are changes.
+#[cfg(feature = "json-models")]
 pub mod ship_locker;
 
 /// Provides some utility functions and macros that are used internally.
