--- src/modules/exploration/functions.rs
@@ -1,1 +1,0 @@
-pub mod calculate_estimated_worth;
