+++ src/modules/exploration/mod.rs
@@ -1,8 +1,6 @@
-pub use functions::calculate_estimated_worth::calculate_estimated_worth;
 pub use models::codex_entry::CodexEntry;
 pub use models::planetary_signal_type::PlanetarySignalType;
 pub use models::star_class_codex_entry::StarClassCodexEntry;
 pub use models::star_class_codex_entry::StarClassCodexEntryError;
 
-mod functions;
 mod models;
